+++ .gitignore
@@ -7,3 +7,5 @@ __pycache__/
 *.db
 *.sqlite3
 .pytest_cache/
+/dist
+/*.egg-info
